Moving aggregator models to common  - These models were duplicated  in the aggregator and engine
