--- aggregator/src/main/java/com/redhat/cloud/notifications/EmailSubscriptionType.java
@@ -1,5 +1,0 @@
-package com.redhat.cloud.notifications;
-
-public enum EmailSubscriptionType {
-    DAILY
-}
