--- aggregator/src/main/java/com/redhat/cloud/notifications/models/EmailAggregationKey.java
@@ -1,34 +1,0 @@
-package com.redhat.cloud.notifications.models;
-
-import javax.validation.constraints.NotNull;
-
-public class EmailAggregationKey {
-
-    @NotNull
-    private String orgId;
-
-    @NotNull
-    private String bundle;
-
-    @NotNull
-    private String application;
-
-    public EmailAggregationKey(String orgId, String bundle, String application) {
-        this.orgId = orgId;
-        this.bundle = bundle;
-        this.application = application;
-    }
-
-    public String getOrgId() {
-        return orgId;
-    }
-
-    public String getBundle() {
-        return bundle;
-    }
-
-    public String getApplication() {
-        return application;
-    }
-
-}
