+++ common/src/main/java/com/redhat/cloud/notifications/models/AggregationCommand.java
