+++ common/src/main/java/com/redhat/cloud/notifications/models/EmailAggregationKey.java
